Develop to Release merge

+++ backend/src/main/java/com/takecare/backend/auth/DTO/LoginResponseDTO.java
@@ -12,5 +12,5 @@ public class LoginResponseDTO {
     private Integer id;
     private String names;
     private String email;
-    private Integer role;
+    private Byte role;
 }
